chore: fix pre-existing format, lint, type, and clippy errors
Why: `bun run check` was failing at `check:base` due to accumulated drift
across JS/TS and Rust files — prettier format, eslint `require-await` and
`no-unnecessary-type-assertion` violations, TypeScript `.disabled` access
on `HTMLElement`, Rust 1.94.0 `derivable_impls` and `unnecessary_cast`
clippy lints, and `rustfmt` across the workspace.

What:
- JS/TS: prettier reformat (12 files), remove async from sync act/test
  callbacks in use-route-history-nav.test, eslint-disable for legitimate
  setState-in-effect, switch .disabled property access to toBeDisabled()
  matcher in link-previews + paper-form-primitives tests.
- Rust: rustfmt --all across workspace, derive Default for OgImageFetchMode
  enum (clippy::derivable_impls), remove unnecessary i64 cast in og_images
  test helper (clippy::unnecessary_cast).

Note: check:coverage (JS branch 97.96% < 98% threshold) was already
failing at the committed state before these changes — tracked by
WORK-V03-COVERAGE-RESIDUAL in BACKLOG.md.

+++ src-tauri/crates/vault-core/src/archive/history/og_images.rs
@@ -763,12 +763,7 @@ mod tests {
     fn list_urls_for_prefetch_honors_the_limit() {
         let connection = open_test_archive();
         for id in 1..=5 {
-            seed_url(
-                &connection,
-                id,
-                &format!("https://example.com/page/{id}"),
-                (id * 1000) as i64,
-            );
+            seed_url(&connection, id, &format!("https://example.com/page/{id}"), id * 1000);
         }
 
         let two = list_urls_for_prefetch(&connection, 2).unwrap();
@@ -799,10 +794,8 @@ mod tests {
         seed_url(&connection, 2, "https://example.com/uncached-new", 5_000);
         seed_url(&connection, 3, "https://example.com/uncached-mid", 3_000);
         seed_url(&connection, 4, "https://example.com/cached-mid", 2_000);
-        upsert_og_image(&connection, &ok_insert("https://example.com/cached-old", b"x"))
-            .unwrap();
-        upsert_og_image(&connection, &ok_insert("https://example.com/cached-mid", b"y"))
-            .unwrap();
+        upsert_og_image(&connection, &ok_insert("https://example.com/cached-old", b"x")).unwrap();
+        upsert_og_image(&connection, &ok_insert("https://example.com/cached-mid", b"y")).unwrap();
 
         let urls = list_urls_for_prefetch(&connection, 10).unwrap();
         assert_eq!(
